Tagging the HDF5 1.8.4-patch1 release.

+++ test/hyperslab.c
@@ -1234,7 +1234,7 @@ main(int argc, char *argv[])
     printf("\n");
 
     /* Set the random # seed */
-    HDsrandom((unsigned)HDtime(NULL));
+    HDsrandom((unsigned long)HDtime(NULL));
 
     /*
      * Open the library explicitly for thread-safe builds, so per-thread
